Complete Day 1
I tried to optimise the first for performance, by putting it in a hash
table. This avoided nested for loops, and made the algorithm O(n).

Then I immediately ran into an obstacle with the second part - I
couldn't see a better solution than O(n**2) - by adding every combo
of two up, and feeding it to the first solution. When attempting this
I found out about .combination, which basically solved the problem
for me - so I went with that.