Solve Day Four Passport Validation
The trickiest bit in this one was parsing the input and turning it
into an object. Setting values for dynamic property names is common
in JS but seems to be a bit harder in Ruby.

There was a lot of trivial code written for part 2, to validate the
input. I considered making it more OO-like (e.g. classes like
HairColor with valid? methods) but I had already spent quite a long
time on the challenge.