Complete Day 7
I really struggled with this one, as I tried to go down a OO-route (hence
the BagFactory class). I was trying to create bags upfront, with a inner_bags
property that contained the bags it wrapped. This worked okay with with my
unit tests, which had a short list of rules, but it didn't work with the puzzle
input.

I realised I had to avoid trying to create all inner bags upfront, and instead
use more procedural code based on the data structures I made inside BagFactory.

This code is definietly not great, and could use some refactoring but I'm a couple
of days behind on challenges. I might refactor later when I can bring myself to
willingly think of bags again.