Complete Day 8
Another challenging one. I tried to go for more OO in this one, with
various command subclasses and factories to parse inputs.

Took a while to get my head around the second part, at first I thought
you had to change the second-to-last instruction from your run, and went
down that route. It was only later I realised the bad instruction could
have been anywhere in the input, at which point I decided to go for a
brute force approach.

To make it more efficient you could try work out the commands involved in
the halting cycle, but I didn't want to spend loads of time and the input
complex enough to warrant optimising it.