Add Day 11 Part 2 solution
Took OO a bit more extreme, with strategies and rule classes to accomodate
the differences between the first part and the second.

I don't think I'd write this code professionally unless I knew more
flexibility would be required down the line, or had a public API (i.e.
totally changing how things are called resulting in breaking changes)
but it was fun to play about with.