Add Day 12 Part 2 solution
This was a strange one - I had the answer for a while, but I must've
been entering it on the site wrong. My test cases were passing, but I
thought it might've been an issue with larger inputs or a very specific
bug. To try identify the issue, I spent a lot of time refactoring.

The FerryPosition class turned out to be a bit of a hinderance - it
ended up simply message passing to the Vector class. So, once I had a
working solution I ripped it out. I also wrapped the navigations to
make a polymorphic interface for both ship and waypoint positions, but
this just hid the logic under a unneccessary layer.

I think this could've been done cleaner, and certainly shorter, but I'm
trying to improve and build things according to SOLID principles as a
learning exercise. Usually I'm a more YAGNI.