Complete Day 13 Part 2
This was really really tricky, especially after a straightforward part
1. It took me a while to notice they were primes, and then I went on
a Google deep dive to find out about the Chinese Remainder Theorem.
I tried to implement it as is, and it was very tricky to follow.

After a few hours, I found Uncle Bob's video on it:
https://www.youtube.com/watch?v=tHTDAUV-VsM

Whilst trying to avoid spoiling the implementation, his explanation
and solution made a lot more sense to me (as I haven't done raw algorithms
in a long time).

The last issue was a subtle bug that only appeared on the challenge input
 -I was relying on waiting_time_for_next_departure(...) being equal
to the index of the input, to find a suitable timestamp for that bus
departure. However when the bus ID is small (e.g. 13), at an index of
20, the waiting time is never going to be longer than 13 (as you're
doing mod 13), so you'd never get a match.