Add MatchingBusList class to handle the sequential timestamps
I added a lot of loop variables when working on Part 2, as I thought I
had a performance issue with large numbers and wanted to reduce the
number of allocations I was doing (e.g. avoiding .reduce and .select).

However when the issue turned out to be a logic bug, I wanted to go back
to a more functional, immutable style to make it easier to reason about.

I used this opportunity to extract the functionality out into a seperate
class as well.