Add Day 15 Solution
This was interesting. In my first part I only used an array, @sequence, to
track the values. This worked okay, but I was using various filters to
find the 'indexes_of_previous_number', so as my input got larger, the
algorithm got significantly slower (it ended up being O(n^2) complex).

When the second part asked for a much larger number, I knew my algorithm
would choke. So I moved to storing indexes in a hash, to avoid re-looking
through the @sequence array every time I iterated the loop. This brought
the algorithm down to O(n) but it still choked on larger values, as I
was storing every index for every number in an array, and then performing
operations on those arrays.

The final optimisation I did, which made it possible to compute the value
in ~20s, was to only store the last two indexes, so each subsequent
iteration wasn't dealing with larger and larger arrays.